requirements.txt
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
pillow==12.1.0
pyaes==1.6.1
pyasn1==0.6.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-telegram-bot==22.6
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
Telethon==1.42.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
